refactor(web): open the shared web app link in a new tab
The Web App tile in the shared platform-link grid was a react-router
Link, so it navigated in place while the two store tiles already opened
a new tab. Render it as a plain anchor with target="_blank" and drop the
now-unused webRoute prop, since webHref already carries the absolute URL.

Relocate MobileAppQrCode next to the other shared app-link building
blocks so screens outside the review feature can use it.
