+++ apps/web/src/screens/share/MobileAppQrCode.tsx
